reTurn: fix some compile warnings
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/resiprocate-1.8@10419 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ reTurn/AsyncTlsSocketBase.cxx
@@ -36,7 +36,7 @@ AsyncTlsSocketBase::~AsyncTlsSocketBase()
 unsigned int 
 AsyncTlsSocketBase::getSocketDescriptor() 
 { 
-   return mSocket.lowest_layer().native(); 
+   return (unsigned int)mSocket.lowest_layer().native(); 
 }
 
 asio::error_code
